+++ src/pages/public_registration/details/format-apa.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest';
 import { Contributor, ContributorRole } from '../../../types/contributor.types';
 import { JournalType } from '../../../types/publicationFieldNames';
 import { Registration } from '../../../types/registration.types';
-import { formatAPA } from './formatAPA';
+import { formatAPA } from './format-apa';
 
 const buildJournalArticle = (overrides: {
   contributors?: Contributor[];
