+++ src/pages/public_registration/details/format-apa.ts
